mtg_elo_backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
MarkupSafe==3.0.2
psycopg==3.2.9
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyotp==2.9.0
python-decouple==3.8
sqlparse==0.5.3
Werkzeug==3.1.3
